Added functionality for optional __using__ macro for putting ABNF grammar etc in the module calling us:
  * a simple parse("rulename", 'code', %State{}) function calling ABNF.apply passing ourselve as grammar
  * rule("rulename") returning the ABNF AST for that specific rule
  * all inline Elixir functions

For this to work, also had to make changes to grammar for optionally returning the elixir AST to be
stored in the module from which the ABNF __using__ macro is called. Also make it configurable to not
create the module and functions by itself.

Lastly the interpreter needed a change to allow the grammar argument to be passed as Module (atom).

Usage:

  iex> defmodule Basic do use ABNF, grammar_file: "test/resources/basic.abnf" end
  iex> Basic.parse "string1", 'test'
  %ABNF.CaptureResult{input: 'test', rest: [], state: %{}, string_text: 'test',
   string_tokens: ['test'],
    values: [[[[['t']]], [[['e']]], [[['s']]], [[['t']]]]]}